Move all add text files to corresponding module
