+++ features/text/add/build.gradle.kts
@@ -8,4 +8,11 @@ android {
 
 dependencies {
 
+    implementation(project(":shared:designsystem:resources"))
+
+    implementation(project(":features:text:shared"))
+
+    implementation(project(":shared:files:files-logic"))
+    implementation(project(":shared:files:files-data"))
+    implementation(project(":shared:files:files-data"))
 }
