+++ features/text/add/src/main/java/ir/mehdiyari/krypt/features/text/add/AddTextNavigation.kt
@@ -1,4 +1,4 @@
-package ir.mehdiyari.krypt.ui.text.add
+package ir.mehdiyari.krypt.features.text.add
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.ui.Modifier
