+++ features/text/add/src/main/res/values/strings.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <string name="text_title">Title</string>
+    <string name="text_content">Content</string>
+    <string name="delete_text">Delete</string>
+    <string name="save_text">Save</string>
+
+    <string name="title_must_not_empty">Title must not be empty</string>
+    <string name="content_must_not_empty">Content must not be empty</string>
+
+    <string name="cant_find_these_text">We can\'t find these text</string>
+    <string name="error_while_decrypting">Error while decrypting text file</string>
+
+    <string name="successfully_encrypt_note">Your note was successfully encrypted</string>
+    <string name="failed_to_encrypt_note">We failed to encrypt your note</string>
+
+    <string name="delete_text_was_successfully">Text file successfully deleted</string>
+    <string name="delete_text_was_unsuccessfully">Error while removing text file</string>
+</resources>
